feat : create Chat - ChatListView - Debouncer
